Update Makefile and GHA Dev1 Workflow
Why these changes are being introduced:
After some initial testing by the developer, we are moving the control
of the ECR Repository for this function into the hands of Infra through
the mitlib-tf-workloads-ecr repository. That repository generates the
correct Makefile commands for the developer to push containers to the
managed ECR Repository and update the associated Lambda function.

How this addresses that need:
* Update the Makefile with the output from mitlib-tf-workloads-ecr
* Create a new dev-build.yml GitHub Actions workflow using the output
from mitlib-tf-workloads-ecr

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1011