Use Flask-Login for user session management
Why these changes are being introduced:
* The Application Load Balancer (ALB) that manages
authentication and authorization provides an access token
and user claims (info about user) in the request headers
routed to its target, the Flask app. With Flask-Login,
this data can be used to identify the 'current_user'
on the app and detect when a session may be stale.

How this addresses that need:
* Implement 'request_loader' callback method
* Add utility method to parse OIDC data from request headers
* Add 'User' class
* Add 'logout' view
* Decorate views using 'login_required'
* Add 'AWS_DEFAULT_REGION' to Config
* Add unit tests

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1033