Add option to ignore "login_required"
Why these changes are being introduced:
* The application configuration variable LOGIN_DISABLED
is a convenient method to globally turn off the login method
(i.e., parsing OIDC data from header) for purposes of
unit testing and running the app in Dev1 (ALB is turned off for Dev1).

How this addresses that need:
* Add environment variable 'LOGIN_DISABLED' to Config

Side effects of this change:
* Environment variable LOGIN_DISABLED must be set to "true" for Dev1
and "false" for Prod and Stage.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1033