Log actions taken by users
Why these changes are being introduced:
* Monitor actions taken by users in the Flask
app to identify who executed a task run. This
information will be useful when debugging /
troubleshooting an error, as well as help
us to maintain an audit of invoice processing runs.

How this addresses that need:
* Add utility method for logging activity
* Add logs to track executed runs, status checks, and logouts
* Update homepage to welcome logged in user

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1034